refactor(animations): igniteui-angular/animations
