+++ projects/igniteui-angular/animations/README.md
