+++ projects/igniteui-angular/animations/ng-package.json
@@ -1,0 +1,5 @@
+{
+  "lib": {
+    "entryFile": "src/main.ts"
+  }
+}
