+++ projects/igniteui-angular/animations/src/fade/README.md
@@ -21,7 +21,7 @@ const params: IAnimationParams = {
 If parameters are attached, they act as default values.  When an animation is invoked via [`useAnimation`](https://angular.io/api/animations/useAnimation) then parameter values are allowed to be passed in directly. If any of the passed in parameter values are missing then the default values will be used.
 
 ``` typescript
-import { fadeIn } from "igniteui-angular/animations/main";
+import { fadeIn } from "igniteui-angular/animations";
 import { EaseOut } from "ignieui-angular/animations/easings";
 
 useAnimation(fadeIn, {
