+++ projects/igniteui-angular/animations/src/flip/README.md
@@ -32,7 +32,7 @@ const params: IAnimationParams = {
 If parameters are attached, they act as default values.  When an animation is invoked via [`useAnimation`](https://angular.io/api/animations/useAnimation) then parameter values are allowed to be passed in directly. If any of the passed in parameter values are missing then the default values will be used.
 
 ``` typescript
-import { flipTop } from "igniteui-angular/animations/main";
+import { flipTop } from "igniteui-angular/animations";
 
 useAnimation(fadeIn);
 ```
