+++ projects/igniteui-angular/animations/src/grow/index.ts
