+++ projects/igniteui-angular/animations/src/interface.ts
