+++ projects/igniteui-angular/animations/src/main.ts
@@ -1,0 +1,19 @@
+export { IAnimationParams } from './interface';
+export { AnimationUtil } from './util';
+export { fadeIn, fadeOut } from './fade/index';
+export {
+    flipTop,
+    flipRight,
+    flipBottom,
+    flipLeft,
+    flipHorFwd,
+    flipHorBck,
+    flipVerFwd,
+    flipVerBck
+} from './flip/index';
+export * from './rotate/index';
+export * from './misc/index';
+export * from './scale/index';
+export * from './slide/index';
+export * from './swing/index';
+export * from './grow/index';
