+++ projects/igniteui-angular/animations/src/misc/index.ts
@@ -1,0 +1,14 @@
+export {
+    shakeHor,
+    shakeVer,
+    shakeTop,
+    shakeBottom,
+    shakeRight,
+    shakeLeft,
+    shakeCenter,
+    shakeTr,
+    shakeBr,
+    shakeBl,
+    shakeTl
+} from './shake';
+export { pulsateFwd, pulsateBck, heartbeat, blink } from './pulsate';
