+++ projects/igniteui-angular/animations/src/rotate/README.md
@@ -52,7 +52,7 @@ const params: IAnimationParams = {
 If parameters are attached, they act as default values.  When an animation is invoked via [`useAnimation`](https://angular.io/api/animations/useAnimation) then parameter values are allowed to be passed in directly. If any of the passed in parameter values are missing then the default values will be used.
 
 ``` typescript
-import { rotateInCenter } from "igniteui-angular/animations/main";
+import { rotateInCenter } from "igniteui-angular/animations";
 
 useAnimation(rotateInCenter);
 ```
