+++ projects/igniteui-angular/animations/src/rotate/index.ts
