+++ projects/igniteui-angular/animations/src/scale/index.ts
