+++ projects/igniteui-angular/animations/src/slide/index.ts
