+++ projects/igniteui-angular/animations/src/swing/index.ts
