--- projects/igniteui-angular/src/lib/animations/main.ts
@@ -1,9 +1,0 @@
-export * from './interface';
-export * from './fade/index';
-export * from './flip/index';
-export * from './rotate/index';
-export * from './misc/index';
-export * from './scale/index';
-export * from './slide/index';
-export * from './swing/index';
-export * from './grow/index';
