--- projects/igniteui-angular/src/lib/animations/misc/index.ts
@@ -1,2 +1,0 @@
-export * from './shake';
-export * from './pulsate';
