+++ projects/igniteui-angular/src/lib/calendar/calendar.component.ts
@@ -13,8 +13,9 @@ import {
     QueryList,
     OnDestroy
 } from '@angular/core';
+import { NgIf, NgTemplateOutlet, NgStyle, NgFor, DatePipe } from '@angular/common';
 import { NG_VALUE_ACCESSOR } from '@angular/forms';
-import { fadeIn, scaleInCenter, slideInLeft, slideInRight } from '../animations/main';
+
 import { IgxCalendarHeaderTemplateDirective, IgxCalendarSubheaderTemplateDirective, IgxCalendarScrollMonthDirective } from './calendar.directives';
 import { ICalendarDate, IgxCalendarView, ScrollMonth, monthRange } from './calendar';
 import { IgxMonthPickerBaseDirective } from './month-picker/month-picker-base';
@@ -26,7 +27,7 @@ import { takeUntil, debounce, skipLast, switchMap } from 'rxjs/operators';
 import { IViewChangingEventArgs } from './days-view/days-view.interface';
 import { IgxMonthViewSlotsCalendar, IgxGetViewDateCalendar } from './months-view.pipe';
 import { IgxIconComponent } from '../icon/icon.component';
-import { NgIf, NgTemplateOutlet, NgStyle, NgFor, DatePipe } from '@angular/common';
+import { fadeIn, scaleInCenter, slideInLeft, slideInRight } from 'igniteui-angular/animations';
 
 let NEXT_ID = 0;
