+++ projects/igniteui-angular/src/lib/carousel/carousel-base.ts
@@ -1,10 +1,9 @@
 import { AnimationReferenceMetadata, useAnimation } from '@angular/animations';
 import { ChangeDetectorRef, EventEmitter, Inject } from '@angular/core';
-import { fadeIn } from '../animations/fade';
-import { slideInLeft } from '../animations/slide';
 import { mkenum } from '../core/utils';
 import { IgxAngularAnimationService } from '../services/animation/angular-animation-service';
 import { AnimationPlayer, AnimationService } from '../services/animation/animation';
+import { fadeIn, slideInLeft } from 'igniteui-angular/animations';
 
 export enum Direction { NONE, NEXT, PREV }
 
@@ -102,9 +101,9 @@ export abstract class IgxCarouselComponentBase {
             duration = this.animationPosition ? this.animationPosition * this.defaultAnimationDuration : this.defaultAnimationDuration;
         }
 
+        const trans = this.animationPosition ? this.animationPosition * 100 : 100;
         switch (this.animationType) {
             case HorizontalAnimationType.slide:
-                const trans = this.animationPosition ? this.animationPosition * 100 : 100;
                 return {
                     enterAnimation: useAnimation(slideInLeft,
                         {
