+++ projects/igniteui-angular/src/lib/core/utils.ts
@@ -1,28 +1,7 @@
-import { AnimationReferenceMetadata } from '@angular/animations';
 import { CurrencyPipe, formatDate as _formatDate, isPlatformBrowser } from '@angular/common';
 import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { mergeWith } from 'lodash-es';
 import { Observable } from 'rxjs';
-import {
-    blink, fadeIn, fadeOut, flipBottom, flipHorBck, flipHorFwd, flipLeft, flipRight, flipTop,
-    flipVerBck, flipVerFwd, growVerIn, growVerOut, heartbeat, pulsateBck, pulsateFwd, rotateInBl,
-    rotateInBottom, rotateInBr, rotateInCenter, rotateInDiagonal1, rotateInDiagonal2, rotateInHor,
-    rotateInLeft, rotateInRight, rotateInTl, rotateInTop, rotateInTr, rotateInVer, rotateOutBl,
-    rotateOutBottom, rotateOutBr, rotateOutCenter, rotateOutDiagonal1, rotateOutDiagonal2,
-    rotateOutHor, rotateOutLeft, rotateOutRight, rotateOutTl, rotateOutTop, rotateOutTr,
-    rotateOutVer, scaleInBl, scaleInBottom, scaleInBr, scaleInCenter, scaleInHorCenter,
-    scaleInHorLeft, scaleInHorRight, scaleInLeft, scaleInRight, scaleInTl, scaleInTop, scaleInTr,
-    scaleInVerBottom, scaleInVerCenter, scaleInVerTop, scaleOutBl, scaleOutBottom, scaleOutBr,
-    scaleOutCenter, scaleOutHorCenter, scaleOutHorLeft, scaleOutHorRight, scaleOutLeft,
-    scaleOutRight, scaleOutTl, scaleOutTop, scaleOutTr, scaleOutVerBottom, scaleOutVerCenter,
-    scaleOutVerTop, shakeBl, shakeBottom, shakeBr, shakeCenter, shakeHor, shakeLeft, shakeRight,
-    shakeTl, shakeTop, shakeTr, shakeVer, slideInBl, slideInBottom, slideInBr, slideInLeft,
-    slideInRight, slideInTl, slideInTop, slideInTr, slideOutBl, slideOutBottom, slideOutBr,
-    slideOutLeft, slideOutRight, slideOutTl, slideOutTop, slideOutTr, swingInBottomBck,
-    swingInBottomFwd, swingInLeftBck, swingInLeftFwd, swingInRightBck, swingInRightFwd,
-    swingInTopBck, swingInTopFwd, swingOutBottomBck, swingOutBottomFwd, swingOutLeftBck,
-    swingOutLefttFwd, swingOutRightBck, swingOutRightFwd, swingOutTopBck, swingOutTopFwd
-} from '../animations/main';
 import { setImmediate } from './setImmediate';
 import { isDevMode } from '@angular/core';
 
@@ -578,259 +557,6 @@ export const yieldingLoop = (count: number, chunkSize: number, callback: (index:
 };
 
 export const isConstructor = (ref: any) => typeof ref === 'function' && Boolean(ref.prototype) && Boolean(ref.prototype.constructor);
-
-export class AnimationUtil {
-    private static _instance: AnimationUtil;
-
-    private oppositeAnimation: Map<AnimationReferenceMetadata, AnimationReferenceMetadata> = new Map([
-        [fadeIn, fadeIn],
-        [fadeOut, fadeOut],
-        [flipTop, flipBottom],
-        [flipBottom, flipTop],
-        [flipRight, flipLeft],
-        [flipLeft, flipRight],
-        [flipHorFwd, flipHorBck],
-        [flipHorBck, flipHorFwd],
-        [flipVerFwd, flipVerBck],
-        [flipVerBck, flipVerFwd],
-        [growVerIn, growVerIn],
-        [growVerOut, growVerOut],
-        [heartbeat, heartbeat],
-        [pulsateFwd, pulsateBck],
-        [pulsateBck, pulsateFwd],
-        [blink, blink],
-        [shakeHor, shakeHor],
-        [shakeVer, shakeVer],
-        [shakeTop, shakeTop],
-        [shakeBottom, shakeBottom],
-        [shakeRight, shakeRight],
-        [shakeLeft, shakeLeft],
-        [shakeCenter, shakeCenter],
-        [shakeTr, shakeTr],
-        [shakeBr, shakeBr],
-        [shakeBl, shakeBl],
-        [shakeTl, shakeTl],
-        [rotateInCenter, rotateInCenter],
-        [rotateOutCenter, rotateOutCenter],
-        [rotateInTop, rotateInBottom],
-        [rotateOutTop, rotateOutBottom],
-        [rotateInRight, rotateInLeft],
-        [rotateOutRight, rotateOutLeft],
-        [rotateInLeft, rotateInRight],
-        [rotateOutLeft, rotateOutRight],
-        [rotateInBottom, rotateInTop],
-        [rotateOutBottom, rotateOutTop],
-        [rotateInTr, rotateInBl],
-        [rotateOutTr, rotateOutBl],
-        [rotateInBr, rotateInTl],
-        [rotateOutBr, rotateOutTl],
-        [rotateInBl, rotateInTr],
-        [rotateOutBl, rotateOutTr],
-        [rotateInTl, rotateInBr],
-        [rotateOutTl, rotateOutBr],
-        [rotateInDiagonal1, rotateInDiagonal1],
-        [rotateOutDiagonal1, rotateOutDiagonal1],
-        [rotateInDiagonal2, rotateInDiagonal2],
-        [rotateOutDiagonal2, rotateOutDiagonal2],
-        [rotateInHor, rotateInHor],
-        [rotateOutHor, rotateOutHor],
-        [rotateInVer, rotateInVer],
-        [rotateOutVer, rotateOutVer],
-        [scaleInTop, scaleInBottom],
-        [scaleOutTop, scaleOutBottom],
-        [scaleInRight, scaleInLeft],
-        [scaleOutRight, scaleOutLeft],
-        [scaleInBottom, scaleInTop],
-        [scaleOutBottom, scaleOutTop],
-        [scaleInLeft, scaleInRight],
-        [scaleOutLeft, scaleOutRight],
-        [scaleInCenter, scaleInCenter],
-        [scaleOutCenter, scaleOutCenter],
-        [scaleInTr, scaleInBl],
-        [scaleOutTr, scaleOutBl],
-        [scaleInBr, scaleInTl],
-        [scaleOutBr, scaleOutTl],
-        [scaleInBl, scaleInTr],
-        [scaleOutBl, scaleOutTr],
-        [scaleInTl, scaleInBr],
-        [scaleOutTl, scaleOutBr],
-        [scaleInVerTop, scaleInVerBottom],
-        [scaleOutVerTop, scaleOutVerBottom],
-        [scaleInVerBottom, scaleInVerTop],
-        [scaleOutVerBottom, scaleOutVerTop],
-        [scaleInVerCenter, scaleInVerCenter],
-        [scaleOutVerCenter, scaleOutVerCenter],
-        [scaleInHorCenter, scaleInHorCenter],
-        [scaleOutHorCenter, scaleOutHorCenter],
-        [scaleInHorLeft, scaleInHorRight],
-        [scaleOutHorLeft, scaleOutHorRight],
-        [scaleInHorRight, scaleInHorLeft],
-        [scaleOutHorRight, scaleOutHorLeft],
-        [slideInTop, slideInBottom],
-        [slideOutTop, slideOutBottom],
-        [slideInRight, slideInLeft],
-        [slideOutRight, slideOutLeft],
-        [slideInBottom, slideInTop],
-        [slideOutBottom, slideOutTop],
-        [slideInLeft, slideInRight],
-        [slideOutLeft, slideOutRight],
-        [slideInTr, slideInBl],
-        [slideOutTr, slideOutBl],
-        [slideInBr, slideInTl],
-        [slideOutBr, slideOutTl],
-        [slideInBl, slideInTr],
-        [slideOutBl, slideOutTr],
-        [slideInTl, slideInBr],
-        [slideOutTl, slideOutBr],
-        [swingInTopFwd, swingInBottomFwd],
-        [swingOutTopFwd, swingOutBottomFwd],
-        [swingInRightFwd, swingInLeftFwd],
-        [swingOutRightFwd, swingOutLefttFwd],
-        [swingInLeftFwd, swingInRightFwd],
-        [swingOutLefttFwd, swingOutRightFwd],
-        [swingInBottomFwd, swingInTopFwd],
-        [swingOutBottomFwd, swingOutTopFwd],
-        [swingInTopBck, swingInBottomBck],
-        [swingOutTopBck, swingOutBottomBck],
-        [swingInRightBck, swingInLeftBck],
-        [swingOutRightBck, swingOutLeftBck],
-        [swingInBottomBck, swingInTopBck],
-        [swingOutBottomBck, swingOutTopBck],
-        [swingInLeftBck, swingInRightBck],
-        [swingOutLeftBck, swingOutRightBck],
-    ]);
-
-    private horizontalAnimations: AnimationReferenceMetadata[] = [
-        flipRight,
-        flipLeft,
-        flipVerFwd,
-        flipVerBck,
-        rotateInRight,
-        rotateOutRight,
-        rotateInLeft,
-        rotateOutLeft,
-        rotateInTr,
-        rotateOutTr,
-        rotateInBr,
-        rotateOutBr,
-        rotateInBl,
-        rotateOutBl,
-        rotateInTl,
-        rotateOutTl,
-        scaleInRight,
-        scaleOutRight,
-        scaleInLeft,
-        scaleOutLeft,
-        scaleInTr,
-        scaleOutTr,
-        scaleInBr,
-        scaleOutBr,
-        scaleInBl,
-        scaleOutBl,
-        scaleInTl,
-        scaleOutTl,
-        scaleInHorLeft,
-        scaleOutHorLeft,
-        scaleInHorRight,
-        scaleOutHorRight,
-        slideInRight,
-        slideOutRight,
-        slideInLeft,
-        slideOutLeft,
-        slideInTr,
-        slideOutTr,
-        slideInBr,
-        slideOutBr,
-        slideInBl,
-        slideOutBl,
-        slideInTl,
-        slideOutTl,
-        swingInRightFwd,
-        swingOutRightFwd,
-        swingInLeftFwd,
-        swingOutLefttFwd,
-        swingInRightBck,
-        swingOutRightBck,
-        swingInLeftBck,
-        swingOutLeftBck,
-    ];
-
-    private  verticalAnimations: AnimationReferenceMetadata[] = [
-        flipTop,
-        flipBottom,
-        flipHorFwd,
-        flipHorBck,
-        growVerIn,
-        growVerOut,
-        rotateInTop,
-        rotateOutTop,
-        rotateInBottom,
-        rotateOutBottom,
-        rotateInTr,
-        rotateOutTr,
-        rotateInBr,
-        rotateOutBr,
-        rotateInBl,
-        rotateOutBl,
-        rotateInTl,
-        rotateOutTl,
-        scaleInTop,
-        scaleOutTop,
-        scaleInBottom,
-        scaleOutBottom,
-        scaleInTr,
-        scaleOutTr,
-        scaleInBr,
-        scaleOutBr,
-        scaleInBl,
-        scaleOutBl,
-        scaleInTl,
-        scaleOutTl,
-        scaleInVerTop,
-        scaleOutVerTop,
-        scaleInVerBottom,
-        scaleOutVerBottom,
-        slideInTop,
-        slideOutTop,
-        slideInBottom,
-        slideOutBottom,
-        slideInTr,
-        slideOutTr,
-        slideInBr,
-        slideOutBr,
-        slideInBl,
-        slideOutBl,
-        slideInTl,
-        slideOutTl,
-        swingInTopFwd,
-        swingOutTopFwd,
-        swingInBottomFwd,
-        swingOutBottomFwd,
-        swingInTopBck,
-        swingOutTopBck,
-        swingInBottomBck,
-        swingOutBottomBck,
-    ];
-
-    private constructor() { }
-
-    public static instance() {
-        return this._instance || (this._instance = new this());
-    }
-
-    public reverseAnimationResolver(animation: AnimationReferenceMetadata): AnimationReferenceMetadata {
-        return this.oppositeAnimation.get(animation) ?? animation;
-    }
-
-
-    public isHorizontalAnimation(animation: AnimationReferenceMetadata): boolean {
-        return this.horizontalAnimations.includes(animation);
-    }
-
-    public isVerticalAnimation(animation: AnimationReferenceMetadata): boolean {
-        return this.verticalAnimations.includes(animation);
-    }
-}
 
 /**
  * Similar to Angular's formatDate. However it will not throw on `undefined | null | ''` instead
