+++ projects/igniteui-angular/src/lib/grids/grid.common.ts
@@ -1,8 +1,8 @@
 import { Directive } from '@angular/core';
 import { ConnectedPositioningStrategy } from '../services/public_api';
 import { VerticalAlignment, PositionSettings, Point } from '../services/overlay/utilities';
-import { scaleInVerBottom, scaleInVerTop } from '../animations/main';
 import { IgxForOfSyncService } from '../directives/for-of/for_of.sync.service';
+import { scaleInVerBottom, scaleInVerTop } from 'igniteui-angular/animations';
 
 
 @Directive({
