+++ projects/igniteui-angular/src/lib/services/overlay/overlay.ts
@@ -17,23 +17,7 @@ import {
 } from '@angular/core';
 import { fromEvent, Subject, Subscription } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
-import {
-    fadeIn,
-    fadeOut,
-    IAnimationParams,
-    scaleInHorLeft,
-    scaleInHorRight,
-    scaleInVerBottom,
-    scaleInVerTop,
-    scaleOutHorLeft,
-    scaleOutHorRight,
-    scaleOutVerBottom,
-    scaleOutVerTop,
-    slideInBottom,
-    slideInTop,
-    slideOutBottom,
-    slideOutTop
-} from '../../animations/main';
+
 import { PlatformUtil } from '../../core/utils';
 import { IgxOverlayOutletDirective } from '../../directives/toggle/toggle.directive';
 import { IgxAngularAnimationService } from '../animation/angular-animation-service';
@@ -60,6 +44,7 @@ import {
     RelativePositionStrategy,
     VerticalAlignment
 } from './utilities';
+import { fadeIn, fadeOut, IAnimationParams, scaleInHorLeft, scaleInHorRight, scaleInVerBottom, scaleInVerTop, scaleOutHorLeft, scaleOutHorRight, scaleOutVerBottom, scaleOutVerTop, slideInBottom, slideInTop, slideOutBottom, slideOutTop } from 'igniteui-angular/animations';
 
 /**
  * [Documentation](https://www.infragistics.com/products/ignite-ui-angular/angular/components/overlay-main)
