+++ projects/igniteui-angular/src/lib/services/overlay/position/auto-position-strategy.ts
@@ -1,7 +1,7 @@
 import { AnimationReferenceMetadata } from '@angular/animations';
-import { AnimationUtil } from '../../../core/utils';
 import { ConnectedFit, HorizontalAlignment, VerticalAlignment } from './../utilities';
 import { BaseFitPositionStrategy } from './base-fit-position-strategy';
+import { AnimationUtil } from 'igniteui-angular/animations';
 
 /**
  * Positions the element as in **Connected** positioning strategy and re-positions the element in
