+++ projects/igniteui-angular/src/lib/toast/toast.component.ts
@@ -22,7 +22,7 @@ import {
 import { IgxNotificationsDirective } from '../directives/notification/notifications.directive';
 import { ToggleViewEventArgs } from '../directives/toggle/toggle.directive';
 import { useAnimation } from '@angular/animations';
-import { fadeIn, fadeOut } from '../animations/fade';
+import { fadeIn, fadeOut } from 'igniteui-angular/animations';
 
 let NEXT_ID = 0;
