+++ tsconfig.json
@@ -22,6 +22,9 @@
         "igniteui-angular": [
           "projects/igniteui-angular/src/public_api.ts"
         ],
+        "igniteui-angular/*": [
+          "projects/igniteui-angular/*"
+        ],
         "igniteui-angular-i18n": [
           "projects/igniteui-angular-i18n/src/index.ts"
         ]
